+++ bin/snowflaker.dart
@@ -1,14 +1,14 @@
 import 'dart:async';
 
-import 'package:snowflake/snowflake.dart';
+import 'package:snowflaker/snowflaker.dart';
 
 void main(List<String> arguments) {
-  final snowflake = Snowflake(
+  final snowflaker = Snoflaker(
     workerId: 1,
     datacenterId: 1,
   );
-  final id = snowflake.nextId();
-  final id2 = snowflake.nextId();
+  final id = snowflaker.nextId();
+  final id2 = snowflaker.nextId();
   Zone.current.print('id: $id');
   Zone.current.print('id2: $id2');
 }